// src/app/models/flight.ts
import {Airport} from './airport';

export class Flight {
  date: string;
  airline: string;
  nameAirporOrigin: string;
  descAirporOrigin: string;
  iataAirporOrigin: string;
  cityNameAirporOrigin: string;
  cityDescAirporOrigin: string;

  nameAirporDestination: string;
  descAirporDestination: string;
  iataAirporDestination: string;
  cityNameAirporDestination: string;
  cityDescAirporDestination: string;
}